+++ app/svg-artifact-generator/page.tsx
@@ -1,0 +1,5 @@
+import SvgArtifactGeneratorPanel from "../../components/SvgArtifactGeneratorPanel";
+
+export default function SvgArtifactGeneratorPage() {
+  return <SvgArtifactGeneratorPanel />;
+}